+++ modules/agent_toolkit_cli/agent_toolkit_cli_test.v
@@ -1,5 +1,5 @@
 module agent_toolkit_cli
 
-fn test_run() {
-	assert run() == 'ok'
+fn test_ping_core() {
+	assert ping_core() == 'ok'
 }